apps/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
httpx==0.27.2
pytest==8.3.3
ruff==0.7.3
